make bat/exa optional installs with EXTRA env variable

+++ system/install.sh
@@ -4,10 +4,13 @@ SRC_DIR=$(dirname "$0")
 # System tools / utilities
 if [ "$(uname -s)" = "Darwin" ]
 then
-  brew install exa
   brew install fzf
   brew install htop
-  brew install bat
+  # if env variable extra set
+  if [ -n "$EXTRA" ];    
+    brew install bat
+    brew install exa
+  fi
   $(brew --prefix)/opt/fzf/install --all
 else
   if [ -z "$NO_ROOT" ]; then
@@ -16,14 +19,16 @@ else
   fi
 
   # Exa
-  if ! $(which $MY_BIN_LOC/exa); then
-      (
-          tmp_dir=$(mktemp -d -t exa-XXXXXXX) && cd $tmp_dir
-          wget https://github.com/ogham/exa/releases/download/v0.9.0/exa-linux-x86_64-0.9.0.zip
-          unzip exa-linux-x86_64-0.9.0.zip && chmod a+x exa-linux-x86_64
-          cp exa-linux-x86_64 $MY_BIN_LOC/exa
-          rm -rf $tmp_dir
-      )
+  if [ -n "$EXTRA" ];    
+    if ! $(which $MY_BIN_LOC/exa); then
+        (
+            tmp_dir=$(mktemp -d -t exa-XXXXXXX) && cd $tmp_dir
+            wget https://github.com/ogham/exa/releases/download/v0.9.0/exa-linux-x86_64-0.9.0.zip
+            unzip exa-linux-x86_64-0.9.0.zip && chmod a+x exa-linux-x86_64
+            cp exa-linux-x86_64 $MY_BIN_LOC/exa
+            rm -rf $tmp_dir
+        )
+    fi
   fi
 
   # fd
